feat: ReviewDetail, PaymentsDetail 레이아웃 구성
