--- app/(main)/userInfo/assigned/page.tsx
@@ -1,5 +1,0 @@
-function AssignedMatchPage() {
-  return <div>Assigned Match</div>;
-}
-
-export default AssignedMatchPage;
