--- components/common/MatchItem/MatchItem.test.tsx
@@ -1,9 +1,0 @@
-import { render } from "@testing-library/react";
-
-import MatchItem from "./MatchItem";
-
-describe("MatchItem test", () => {
-  test("MatchItem 렌더링 테스트", () => {
-    // 기능 완성 후, 작성 예정
-  });
-});
