--- components/paymentsDetailPage/history.tsx
@@ -1,5 +1,0 @@
-function History() {
-  return <div>결제 내역 - detail</div>;
-}
-
-export default History;
