--- components/paymentsDetailPage/payments.tsx
@@ -1,5 +1,0 @@
-function Payments() {
-  return <div>결제 페이지</div>;
-}
-
-export default Payments;
