--- components/reservationDetailPage/History.tsx
@@ -1,5 +1,0 @@
-function History() {
-  return <div>예약 내역 detail - 취소, 변경</div>;
-}
-
-export default History;
