--- components/reviewDetailPage/ReviewDetail.tsx
@@ -1,13 +1,0 @@
-import Results from "../common/SearchByPeriod/Results";
-import SearchByPeriod from "../common/SearchByPeriod/SearchByPeriod";
-
-function ReviewDetail() {
-  return (
-    <div>
-      <SearchByPeriod />
-      <Results>검색결과</Results>
-    </div>
-  );
-}
-
-export default ReviewDetail;
